pyproject.toml
[project]
name = "liscrape"
version = "0.1.0"
description = "LinkedIn profile scraping tool using Proxycurl API with Airtable integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.0",
    "pyairtable>=2.3.3",
    "pytest>=8.4.1",
    "openai>=1.97.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
